requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.1.4
django-autoslug
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pyscopg2==2.9.10
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
django-simple-history==3.7.0
django-jazzmin